requirements.txt
rich
mysql-connector-python
